pyproject.toml
[project]
authors = [
  {name = "I Love Study", email = "ilstudy@qq.com"},
]
dependencies = [
    "rich>=13.7.1",
    "rich-argparse>=1.4.0",
    "ffmpeg-progress-yield>=0.7.8",
    "skia-python>=128.0b9",
    "numpy>=2.1.0",
]
description = "Default template for PDM package"
license = {text = "LGPL-3.0"}
name = "py-rotaeno-stablizer-gui"
readme = "README.md"
requires-python = ">=3.11"
version = "0.2.0"

[project.optional-dependencies]
gui = [
    "ttkbootstrap>=1.10.1",
]

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "viztracer>=0.16.2", # performance
    "nuitka>=2.1.3", # packaging
    "yapf>=0.40.2", # formatter
]

[[tool.pdm.autoexport]]
filename = "requirements.txt"
groups = ["default"]
without-hashes = true
